Refactor Cache code to be more readable
